Refactored factory functions
